[Email Reminders] Exclude TN pilot districts, recently logged-in users from emails (Recidiviz/recidiviz-data#51238)
## Description of the change

Changes to the Apps Script for October's email reminders, which
successfully ran on 10/23 and 10/24:
- we will stop emailing users in TN who are in the compliant reporting
2025 policy pilot districts, because we would be sending them incorrect
numbers based on the old CR policy
- we will not email anyone who has logged in within the past 2 days,
because recent user actions that might change eligibility numbers (like
marking people ) would not be reflected in the BQ views that we get
numbers from for the emails.

There was also a request to change the Iowa text from "eligible" that
I'm ignoring for this month because it was just a one-off.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-dashboardsRecidiviz/recidiviz-data#9763

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7bf98fa225fc3115aba95be40c71308a410754a8