[US_NC] Create notebook to perform analyses on Edovo survey responses (Recidiviz/recidiviz-data#50087)
## Description of the change
Adds a notebook which analyzes NC survey data. Claude helped a lot with
generating this code -- it's messy, but @hugosr-r has helped look over
it and we feel okay about it from a statistical perspective (Hugo, let
me know if you disagree!). My ask is that we lean heavily towards
merging this as is, adding todos if there if needed.

There's a few main parts:
1. The notebook (ipynb file)
2. FacilityOutlierAnalyzer: class which does the calculations and math
3. FacilityOutlierReporter: class which generates markdown report and
charts

We've had a few convos about this
[here](https://recidiviz.slack.com/archives/C085XCNQ0TC/p1759859961147269)
and
[here](https://recidiviz.slack.com/archives/C04G37PRQ10/p1759516765358559)
(this ones a bit outdated -- we no longer use IQR, for instance, at
Hugo's suggestion).

I don't commit the input or output to avoid leaking PII, but ask if you
want to have a look!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ad81da6f3b280d43a6543022706d4067e9c77dc8