[Cleanup] Remove gender/prioritized_race_or_ethnicity cols from VIOLATION event observation view (Recidiviz/recidiviz-data#51728)
## Description of the change

Removes the `gender` and `prioritized_race_or_ethnicity` cols from the
observation view for `EventType.VIOLATION`. These were only referenced
in a very old notebook of @samanthahuff's and is the last holdout
reference to these fields on any dataflow metric output table. I think
it makes most sense to delete these and if we really need to do an
analysis with these fields on the VIOLATION event, we can either a) use
this as motivation to address Recidiviz/recidiviz-data#27010 or b) add them back to the event,
pulling from `sessions.person_demographics`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#25565
Part of Recidiviz/recidiviz-data#50972

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b5b9169523061e0d390d0b04b35e91dda93fb3dc