[US_TN] Sentence V2 Migration - Preprocessing View for TN Sentence Statuses (Recidiviz/recidiviz-data#51118)
## Description of the change
This PR creates a preprocessing view that captures TN sentence logic
that has been handled at the criteria level. By moving this to a
preprocessing view, we can utilize our standard state-agnostic approach
to determining active sentences off of sentence_sessions views. The
logic that is replicated is based off of what is in
[this](https://github.com/Recidiviz/recidiviz-data/blob/e5743fc6f21db2dce0fc55f58545b3f2eec96dd7/recidiviz/task_eligibility/criteria/state_specific/us_tn/not_serving_ineligible_cr_offense.py#L52)
criteria (among others). Essentially, when there is no completion date,
we use the sentence projected completion date max as the completion
date. This means we assume that a sentence is no longer active once its
projected max completion date has passed.

This PR also migrates 5 TN-specific criteria off of v1 sentence infra,
which is possible following the creation of this preprocessing view

**Validation**
I could not get a load_views_to_sandbox run to work in auto mode (can
try that again later). In the meantime, I constructed the TN criteria
dependency graph and did a manual run of all relevant criteria and TES
(see validation notebook
[here](https://colab.research.google.com/drive/1zGsg5rPofw_s3miVAGP92bhkX8dvMxVC?authuser=0#scrollTo=xK3dgQUZ2JGa)).

I checked the current population for each of the TN tasks and compared
the eligible and almost eligible population counts. Additionally, I
compared the reasons json to make sure that matched as well.

The result of this validation shows only 1 single discrepancy (a person
moved from being eligible to not eligible for the
TRANSFER_TO_COMPLIANT_REPORTING_NO_DISCRETION task). This is due to the
US_TN_INELIGIBLE_OFFENSES_EXPIRED criteria, which now shows this 1
additional person has having ineligible offenses. The reason for this is
explainable, and I think correct? That criteria (unlike other TN
criteria) **did not** use similar logic that takes into account the
projected completion date max. Therefore, when we do start taking it
into account it caused this person to show up as ineligible.
@mhilderbran, I'd love to chat through the specific example with you in
a bit more detail and make sure we are aligned that the change due to
that criteria is correct.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Andrew Gaidus <agaidus@Andrews-MacBook-Pro.local>
Co-authored-by: Emily Tiry <7997357+etiry@users.noreply.github.com>
Co-authored-by: Emily Turner <149205316+emilyemilyemilyemilyemilyemily@users.noreply.github.com>
Co-authored-by: Santy Mendoza <55068232+santymendoza@users.noreply.github.com>
GitOrigin-RevId: a7ee7801f3f577d868f7c9b9332138ea00ca9d80