[Ingest] Optimize RunValidations step (Recidiviz/recidiviz-data#51865)
## Description of the change

Updates the `RunValidations` step so that instead of running one "find
uniqueness constraint failures" step per combo of entity class and
specific constraint for that class, we instead run a single
`FindUniqueConstraintFailures` transform, which internally creates
tuples that augment entity "critical dicts" with the constraint name and
runs a single GroupBy across ALL entitites / constraints at once.

Successful MI pipeline run
[here](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-11-05_11_12_02-13826192060927374916;step=;mainTab=JOB_GRAPH;logsSeverity=INFO;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))&authuser=0).
With this optimization, the full pipeline graph displays in theUI, where
before it did not!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Partial resolution for Recidiviz/recidiviz-data#51197 (still don't understand the actual
threshold after which we'll stop being able to see the graph in the UI
again).

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7c6fb7764b7470290b54aec95709c1edef175664