[US_MI] Infer more management level raw scores (Recidiviz/recidiviz-data#51497)
## Description of the change

Infer calculated score to be equal to be equal to management level raw
score in cases where the calculated score is not equal to 7, 15, or 23
and the security classification screen is the latest published version
of the screen (02/12).

We already infer and set the calculated score as the management level
raw score in cases where the management level dominates the confinement
level since we know that calculated score is set as the highest score
value between the management level score and the confinement level
score. However, this PR takes the additional step to set calculated
score as the management level raw score in cases where the calculated
score is not equal to 7, 15, or 23. This is because the numeric value of
a confinement level is set as the lowest value in the range for that
level (i.e. the numeric score of confinement level II would be 7), and
so we know that any calculated score that is not the min value of a
range (so not 7, 15, or 23) must be the actual management score.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f70fabf6e7eda9ce72b93be79572e357dd44029e