[CPA] Add route to admin panel for Case Planning Assistant (Recidiviz/recidiviz-data#51909)
## Description of the change

Add route to admin panel for Case Planning Assistant. This is more/less
a "dummy" route (since we don't use it anywhere in the CPA application),
but is necessary to allow a CPA officer to login and go through all of
the Auth0 post-login actions flow.

I learned this morning (many thanks to @danielsmc & @Catacola) that not
having a `route` impacts a critical piece of the post-login action flows
that writes permissions onto a user's `app_metadata`. This prevented
staff that just had a `cpa_staff` role without any routes from
successfully logging in.

More context: we're just leveraging the existing admin panel
infrastructure to generate and surface a user's `pseudonymizedId` in the
auth token (which is required to connect them to our BQ tables and find
+ surface their list of clients), and using feature variants for
upcoming features. At the moment, we have no use for the `routes`
permissions, but could eventually (somewhere far down the line) if this
application is integrated with the staff app.

<img width="849" height="503" alt="Screenshot 2025-11-06 at 1 15 43 PM"
src="https://github.com/user-attachments/assets/806caf42-1698-467b-b147-430e410b42bc"
/>

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

Co-authored-by: Mahmoud <mahmoud@Mahmouds-MacBook-Pro.local>
GitOrigin-RevId: 81bc07445c688eb884ab5ab58ed6cd71e2e4ea8b