[US_TX] `State staff` Refactor pt.2 (Recidiviz/recidiviz-data#51920)
## Description of the change

In the previous refactor, we didn't handle staff without an email
appropriately. Additionally, we had assumed that there was a primary key
in the data, but in reality, there isn't any PKs in TPOFFICER.

This PR removes that staff external ID and also handles officers without
an email correctly.

Validation notes
[here](https://docs.google.com/document/d/1mjmQZnef8EJHDYy_ZovUiDW-pnHUyUQeHCwCv75n_NU/edit?tab=t.0)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e8d01abdc215f3573735aba37e7c3213f82a4f1e