[Public Dashboard] Pipe sex values through to exported views (Recidiviz/recidiviz-data#51927)
## Description of the change

For all public dashboard views, renames references to `gender` to `sex`
and updates the views to read from the `sex` column in
`person_demographics` instead of `gender. Maintains the OUTPUT column
name of `gender` for all exported views (i.e. view dimensions remain
unchanged).

For the `recidivism_rates_by_cohort_by_year` view I also removed all
US_ID logic (US_ID is not exported and no longer supported in the
frontend as of https://github.com/Recidiviz/public-dashboard/pull/681),
which exposed that `analyst_data.session_cohort_reincarceration` was
only being used for US_ID-specific logic in that view and could be
deleted as part of this change as well.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#50972

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 78130e0540fa881d871684009c5c4393e8285eee