[US_TN] JII App - Ensure that two raw data tables are in sync (Recidiviz/recidiviz-data#51739)
## Description of the change

The TN JII app pulls data from two raw data tables
`OffenderSentenceSummary` and `OffenderCredit`. As it currently stands,
we receive and process the `OffenderSentenceSummary` ~3 hours before the
`OffenderCredit` table. Our view update and front end export tend to
occur in this 3 hour window, which means that we are consistently
showing data that is out of sync.

This PR creates a new view upstream that queries these raw data tables
to determine the max update_date that they both have in common. This
date represents the date that we should use for both of the files.

I then create two preprocessed views off of the raw data that queries
the non-latest versions of these tables, to extract the files that match
this date. The downstream views then point to these preprocessed files
instead of the _latest version of the raw data tables.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Andrew Gaidus <agaidus@Andrews-MacBook-Pro.local>
GitOrigin-RevId: ef8531313fee88e19c4ba595fd4995d9a6d29ed9