[US_IX] Adds logic for IN_CUSTODY clients (Recidiviz/recidiviz-data#52110)
## Description of the change

In this PR, we use information in the field `PhysicalLocationTypeDesc`
to determine whether a client is in custody. We do this so that these
clients are removed from tasks since we have been getting a large number
of tickets concerning this issue. Here is a single
[ticket](https://github.com/Recidiviz/recidiviz-dashboards/issues/10455)
as an example. See more
[here](https://github.com/Recidiviz/recidiviz-data/issues/51534).

I ran a successful end to end sandbox, the prefix is `santy_in`

My validation notes are
[here](https://docs.google.com/document/d/1anuql0Rj9hXGv-5SlGX8PsMjuigtkrcI1_7VlJSXZ-U/edit?tab=t.0)
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2670cc05ea5c550aff1164975a5d689d068775d6